% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree.funs.r
\name{addURL2df}
\alias{addURL2df}
\title{add URLs to one column of a data.frame}
\usage{
addURL2df(df1, textColm, urlColm, fromRow = 1, fName = NULL, ...)
}
\arguments{
\item{df1}{data.frame}

\item{textColm}{int or char; the column to display the text of the URL}

\item{urlColm}{int or char; the column of the URL}

\item{fromRow}{int; the starting row}

\item{fName}{char; output file name}

\item{...}{passed to \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}}}
}
\value{
none by default; see the return of \code{\link[openxlsx:saveWorkbook]{openxlsx::saveWorkbook()}} for
details.
}
\description{
Note package \code{openxlsx} is loaded but not attached in this function.
}
\examples{
# none
}
