% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree.funs.r
\name{ctree2splitInfo}
\alias{ctree2splitInfo}
\title{get info of significant variables in each node of a ctree}
\usage{
ctree2splitInfo(cRe, pCut, nodeInds = partykit::nodeids(cRe, terminal = F))
}
\arguments{
\item{cRe}{object returned from \code{\link[partykit:ctree]{partykit::ctree()}}}

\item{pCut}{numeric; p-val cutoff to be significant}

\item{nodeInds}{int vectors, e.g. \code{1:5} for a tree of five nodes, including leaves}
}
\value{
return a list of following items:
\itemize{
\item \code{pVals} is a list of significant variables at each node of \code{nodeInds} excluding leaves;
\itemize{
\item \code{length(pVals)}: number of internal nodes
\item \code{names(pVals)}: the name of splitting variable at each node
\item each item of \code{pVals} is a numeric vector of p-vals for all variables
whose \code{p-val <= pCut}, ordered by increasing p-vals
}
\item \code{spVars.df} a data.frame of following four columns, one node per row,
\itemize{
\item \code{variable}: the name of splitting variable (\code{spVar})
\item \code{pVal}: the p-val of the \code{spVar}
\item \code{sCut}: the splitting point
\item \code{gtOnRight}: T/F; are observations of greater (\code{gt}: greater than) values
on the right branch of the node? The separation of numeric spVar is always
\code{<=} and \code{>}; if the leaves are listed in increasing \code{y}, \code{gtOnRight==T}
means the \code{spVar} has positive association with \code{y}.
}
\item \code{spVars.str}, a string combining all rows of \code{spVars.df}. rows are
separated by ';' and cells are by ','
}

if cRe has only one root node, i.e. no splitting, NULL is returned
}
\description{
see @return for details. Note that package \code{partykit} is loaded, but not
attached, in this function.
}
\examples{
# none
}
