% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctree.funs.r
\name{rmNZV}
\alias{rmNZV}
\title{remove columns of near zero variance in a data.frame or matrix.}
\usage{
rmNZV(df1, minUniPerc = 0.05, minUniCount = 5)
}
\arguments{
\item{df1}{a data.frame or matrix}

\item{minUniPerc, minUniCount}{criteria to remove columns
unique values are all the values except the most common value, e.g. \verb{1,4,2} in \code{c(1,2,3,3,4)}.
\code{uniCount} and \code{uniPerc} are the count and percentage of samples having the unique values
failure to match either leads to removal}
}
\value{
a cleaned \code{df1}; data type is kept even if 0 or 1 column remains
}
\description{
It applies to any data type whereas \code{\link[caret:nearZeroVar]{caret::nearZeroVar()}} is only for numeric columns.
}
\examples{
# none
}
